Replaced the original wires, added an initial script to drag wires and connect them
I'm not an artist so I could not get the wires I ripped from the gif to work the way I wanted, I found a white wire set that I could use in the way I was planning to, and since it is white I could be much more efficient with simply using the same wire colored differently. Currently a wire can be dragged across to the other side, this is sort of the MVP for the connection process.