Fixed functionality of wire script, added camera script, and test text
The wire script was modified to make sure that wires could not connect to the same socket, or even to eachother since the collision check was only looking for any collider not belonging to the gameobject of the script. The camera script is a boilerplate that I use for webGL builds, which helps to keep the entire game in view in different screen sizes. The notes and text have been added as placeholders. I added a free font asset to make the text more closely resemble the gif.